% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_extent.R
\name{get_coords_3857}
\alias{get_coords_3857}
\title{Convert degrees coordinates from EPSG 4326 to 3857}
\usage{
get_coords_3857(lat, long)
}
\arguments{
\item{lat}{numeric vecotr - degrees latitude of point (WGS-84)}

\item{ong}{numeric vecotr - degrees longitude of point (WGS-84)}
}
\description{
A helper function that allows for the calculation of pseudo mercator coords
which are useful when using functions such as `rayshader::render_label()`
because the constructed scene is projected in EPSG 3857.
}
\examples{
.lat <- 57.21956608144513
.long <- -6.092690805001252
get_coords_3857(lat=.lat, long=.long)

}
