% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_3d_vista.R
\name{plot_3d_vista}
\alias{plot_3d_vista}
\title{param}
\usage{
plot_3d_vista(
  lat,
  long,
  radius = 10000,
  elevation_detail = 14,
  overlay_detail = 14,
  zscale = 15,
  img_provider = "Esri.WorldImagery",
  cache_dir = tempdir(),
  ...
)
}
\description{
param
}
